[nasa-itc/deployment#9] Added checks to support Ubuntu local install;

+++ deploy-local-env.sh
@@ -59,12 +59,6 @@ exit 1
 # Version of Ansible to install
 REQ_ANSIBLE_VER="2.9.13"
 
-if [ -f "/etc/debian_version" ]; then
-    EXTRA_VARS="OS=ubuntu GROUND=NONE"
-else
-    EXTRA_VARS="OS=rocky GROUND=NONE"
-fi
-
 ################################################################################
 # PROCESS OPTIONS
 ################################################################################
@@ -103,19 +97,34 @@ fi
 
 ss_announce "Installing Dependencies"
 
-which pip >/dev/null 2>&1
-PIP_INSTALLED=$?
+if [ -f "/etc/debian_version" ]; then
+    EXTRA_VARS="OS=ubuntu GROUND=NONE"
+    which pip >/dev/null 2>&1
+    PIP_INSTALLED=$?
 
-if [[ $PIP_INSTALLED -eq 0 ]]; then
-    ss_status "pip already installed"
+    if [[ $PIP_INSTALLED -eq 0 ]]; then
+    	ss_status "pip already installed"
+    else
+    	sudo apt install -y python3-pip 
+    	sudo pip3 install --upgrade pip
+    	sudo pip3 install markupsafe typing ansible==2.9.13
+    fi
 else
-    # Add repository that hosts package
-    ss_status "Installing pip"
-    sudo yum install -y epel-release
-    sudo yum -y update
-    sudo yum install -y python3-pip
-    sudo pip3 install --upgrade 'pip<21.0'
-    sudo pip3 install markupsafe typing ansible==2.9.13
+    EXTRA_VARS="OS=rocky GROUND=NONE"
+    which pip >/dev/null 2>&1
+    PIP_INSTALLED=$?
+
+    if [[ $PIP_INSTALLED -eq 0 ]]; then
+        ss_status "pip already installed"
+    else
+        # Add repository that hosts package
+        ss_status "Installing pip"
+        sudo yum install -y epel-release
+        sudo yum -y update
+        sudo yum install -y python3-pip
+        sudo pip3 install --upgrade 'pip<21.0'
+        sudo pip3 install markupsafe typing ansible==2.9.13
+    fi
 fi
 
 ####################
